--- src/components/portals/components/Card.jsx
@@ -1,8 +1,0 @@
-import * as React from "react"
-export const Card = () => {
-    return (
-        <div className="content">
-            Left #1
-        </div>
-    )
-}
